test: move mergeAll test to merge-all test file
